Macro for parsing graphql documents at compile time
This addresses issue #33 by adding a LiteralGraphQLStringContext  that adds a graphql method. This allows you to use string interpolation to parse a string literal into a sangria.ast.Document. This is done via a simple macro that calls QueryParse.parse on the string literal pulled from the StringContext call.

The majority of the code is in implementing the Liftable type class for all the ASTNodes. This requires quite a bit of boilerplate, but should be fine for now. Perhaps there is something available that implements this type class automatically for case classes?

This macro allows us to parse the string literal at compile time. Compilation will fail if the string cannot be parsed.