Initialize symbols before checking their annotations.
The Scala compiler uses laziness extensively to deal with recursive types. Annotations are one instance where an uninitialized type will be incorrect, so we call `info` before checking method annotations inside macros. The proper method, initialize, is not available in the public API so I just called .info, which is what most places in the compiler do, anyway.

Fix #172