Add a MappedAbstractType that will transform the input value
- Since Scala 2.x does not have union types, one can represent them a
number of different ways, a common way is to use a boxed type like Either,
or Coproduct from various libraries.  To use these boxed types is impossible
in Sangria because it basically expects the type to unboxed, and
inheriting from `Any` is the only way to achieve that.
- This introduces a concept to map a type `T` to Any such that you can unbox
after the definition, and get a value what the native type defs expect.