Add a faster version of the field selection merging validation
Implements a faster version of OverlappingFieldsCanBeMerged,
that can deal with a high number of fragments or selection sets
without a blowup in runtime.

Also includes JMH for benchmarks.

The implementation uses Java collections as:
- The scala LinkedHashSet implementation is slow
- Java util.ArrayList does not preallocate the array

The anonymous class syntax for consumers and functions is
necessary for compatiblity with scala 2.11, which does not
support SAM lambda syntax yet. It can be refactored to use
lambda, when scala 2.11 support is dropped.

Refs:
https://tech.xing.com/graphql-overlapping-fields-can-be-merged-fast-ea6e92e0a01