always return an empty list for interfaces of interface
The change https://github.com/sangria-graphql/sangria/pull/960
introduces some support for interfaces of interface.

But this feature is not complete: it would need something like
https://github.com/sangria-graphql/sangria/pull/597 for full support.

Right now, sometimes we support interfaces of interface and sometimes now.

To avoid this configuration, let's always return an empty list.

We can tackle the complete support for interfaces of interface later.